// src/common/types/tracking-constants.ts
export enum Actions {
  LOAD = 'load',
  CLICK = 'click',
  ERROR = 'error',
  CLOSE = 'close',

  PRODUCT_VIEW = 'product_view',
  PRODUCT_CLICK = 'product_click',

  RESULT_LOAD = 'result_load',
}

export enum Labels {
  CROP = 'crop',
  FILTER = 'filter',
  SORT = 'sort',
  DOT = 'dot',
  NEW = 'new',
  PAGE = 'page',
  UPLOAD = 'upload',
  SAMPLE = 'sample',
  SEARCH = 'search',
  SEARCH_ERROR = 'search_error',
  CALLOUT = 'callout',
  ENTER = 'enter',
  ICON = 'icon',
}

export enum Category {
  ENTRANCE = 'Entrance',
  UPLOAD = 'Upload',
  RESULT = 'Result',
}